WIP Continue big refactor
