+++ frontend/src/components/editor/editorContext.tsx
@@ -1,21 +1,37 @@
-import { LayerUpdateCallback } from "@/models/editor/layers/layer";
-import { KonvaMouseEvent } from "@/models/editor/utils/events";
+import { LayerId, LayerUpdateCallback } from "@/models/editor/layers/layer";
+import { KonvaMouseEvent, EventType } from "@/models/editor/utils/events";
 import Konva from "konva";
 import { Vector2d } from "konva/lib/types";
-import { createContext, useContext } from "react";
+import React, { createContext, useContext } from "react";
+import { Layer } from "@/models/editor/layers/layer";
+import { VirtualStateSetter } from "./undoRedo";
 
 export type CanvasState = {
   scale: number;
   position: Vector2d;
 };
 
+export type EventHandlers = Partial<Record<EventType, (event?: KonvaMouseEvent) => void>>
+
+// Define the complete context of the editor. This allows child components to edit
+// what is needed in the whole editor.
 interface EditorContextType {
+  isHoldingPrimary: React.RefObject<boolean>;
+  isTransforming: React.RefObject<boolean>;
+
+  updateLayer: (id: LayerId, callback: LayerUpdateCallback, virtual?: boolean) => void;
   editSelectedLayers: (callback: LayerUpdateCallback, virtual?: boolean) => void;
+  commitVirtualLayers: () => void;
+  setVirtualLayers: VirtualStateSetter<Layer[]>;
+
   getCanvasPointerPosition: () => Vector2d;
-  handleLayerSelection: (e: KonvaMouseEvent) => void;
+
   canvasState: CanvasState;
   setCanvasState: React.Dispatch<React.SetStateAction<CanvasState>>;
   stageRef: React.RefObject<Konva.Stage | null>;
+
+  toolEventHandlers: React.RefObject<EventHandlers>;
+  setToolEventHandlers: (eventHandlers: EventHandlers) => void;
 };
 
 export const EditorContext = createContext<EditorContextType | null>(null);
