+++ frontend/src/components/editor/tools/paint.tsx
@@ -15,6 +15,26 @@ export const PaintToolConfigurationComponent = ({
   configuration,
   setConfiguration,
 }: ToolConfigurationProps<PaintToolConfiguration>) => {
+
+  const handleDrawStart = (e: KonvaMouseEvent) => {
+    isDrawing.current = true;
+
+    editSelectedLayers(layer => {
+      const pointPosition = v2Sub(getCanvasPointerPosition(), layer.position);
+      return {
+        ...layer,
+        lines: layer.lines.concat([
+          {
+            points: [pointPosition.x, pointPosition.y],
+            color: "",
+            width: 3,
+            tool: null,
+          }
+        ])
+      };
+    }, true);
+  };
+
   return (
     <div>
       <p className="text-violet-50">
