+++ frontend/src/components/editor/undoRedo.tsx
@@ -1,6 +1,8 @@
 import { useState, useCallback } from "react";
 import { CircularBuffer } from "@/models/editor/utils/CircularBuffer";
 
+export type VirtualStateSetter<T> = (newState: T | ((prev: T) => T)) => void;
+
 /// Hook to store state with undo / redo capabilities.
 /// It has the ability to group multiple smaller modifications into one.
 /// capacity sets the maximum number of states to keep in memory.
